+++ contracts/test/oracles/MockOracle.sol
@@ -17,10 +17,10 @@ contract MockOracle is AbstractOracle {
 
     function stubSetObservation(
         address token,
-        uint256 price,
-        uint256 tokenLiquidity,
-        uint256 quoteTokenLiquidity,
-        uint256 timestamp
+        uint112 price,
+        uint112 tokenLiquidity,
+        uint112 quoteTokenLiquidity,
+        uint32 timestamp
     ) public {
         ObservationLibrary.Observation storage observation = observations[token];
